+++ adapter/in/sdl/components/button/button_size/button_size.h
@@ -18,14 +18,14 @@ typedef enum {
 typedef struct ButtonSize {
     ButtonSizeType size_type;
     Padding padding;
-    int8_t window_percentage; // [0-100]
+    uint8_t window_percentage; // [0-100]
     uint16_t height;
     uint16_t width;
 } ButtonSize;
 
 ButtonSize absolute_button_size(uint16_t height, uint16_t width, Padding padding);
 
-ButtonSize window_relative_button_size(uint16_t window_percentage, Padding padding);
+ButtonSize window_relative_button_size(uint8_t window_percentage, Padding padding);
 
 ButtonSize text_fit_relative_button_size(Padding padding);
